# api/openapi.json
{
  "components": {
    "schemas": {
...
      "AccessToken": {
        "properties": {
          "access_token": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Device": {
        "properties": {
          "browser": {
            "type": "string"
          },
          "device": {
            "type": "string"
          },
          "id": {
            "type": "integer"
          },
          "ip": {
            "type": "string",
            "x-faker": "internet.ip"
          },
          "loginTime": {
            "format": "date-time",
            "type": "string",
            "x-faker": "date.recent"
          },
          "os": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "HTTPError": {
        "properties": {
          "detail": {
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "Login": {
        "properties": {
          "email": {
            "example": "dsx@gmail.com",
            "format": "email",
            "type": "string"
          },
          "password": {
            "example": "Pas$word123",
            "maxLength": 128,
            "minLength": 8,
            "pattern": ".*\\d.*",
            "type": "string"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "type": "object"
      },
...
      "Password": {
        "properties": {
          "password": {
            "example": "Pas$word123",
            "maxLength": 128,
            "minLength": 8,
            "pattern": ".*\\d.*",
            "type": "string"
          }
        },
        "required": [
          "password"
        ],
        "type": "object"
      },
...
      "Register": {
        "properties": {
          "email": {
            "example": "dsx@gmail.com",
            "format": "email",
            "type": "string"
          },
          "firstName": {
            "description": "first name",
            "example": "John",
            "type": "string"
          },
          "lastName": {
            "description": "last name",
            "example": "Doe",
            "type": "string"
          },
          "localityId": {
            "description": "provider's locality id",
            "example": 3167397,
            "type": "integer"
          },
          "localityProvider": {
            "enum": [
              "google",
              "nominatim"
            ],
            "example": "nominatim",
            "type": "string"
          },
          "password": {
            "example": "Pas$word123",
            "maxLength": 128,
            "minLength": 8,
            "pattern": ".*\\d.*",
            "type": "string"
          }
        },
        "required": [
          "email",
          "firstName",
          "lastName",
          "password"
        ],
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "detail": {
            "properties": {
              "<location>": {
                "properties": {
                  "<field_name>": {
                    "items": {
                      "type": "string"
                    },
                    "type": "array"
                  }
                },
                "type": "object"
              }
            },
            "type": "object"
          },
          "message": {
            "type": "string"
          }
        },
        "type": "object"
      },
      "WhoAmI": {
        "properties": {
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "x-faker": "date.past"
          },
          "email": {
            "format": "email",
            "type": "string",
            "x-faker": "internet.email"
          },
          "firstName": {
            "type": "string",
            "x-faker": "name.firstName"
          },
          "id": {
            "type": "integer"
          },
          "isActivated": {
            "type": "boolean"
          },
          "lastName": {
            "type": "string",
            "x-faker": "name.lastName"
          },
          "localityGoogleId": {
            "type": "integer"
          },
          "localityNominatimId": {
            "type": "integer"
          }
        },
        "type": "object"
      }
    },
    "securitySchemes": {
      "csrf_refresh_token": {
        "in": "header",
        "name": "X-CSRF-TOKEN",
        "type": "apiKey"
      },
      "jwt_access_token": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "jwt_refresh_token": {
        "in": "cookie",
        "name": "refresh_token_cookie",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "City Report API",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/auth/activate/{activation_code}": {
      "post": {
        "parameters": [
          {
            "in": "path",
            "name": "activation_code",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "User activated"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Invalid activation code"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Not found"
          }
        },
        "summary": "Activate the user's account",
        "tags": [
          "Authentication operations"
        ]
      }
    },
    "/auth/activation-link-email": {
      "post": {
        "parameters": [],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Activation link sent"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "User already activated"
          }
        },
        "security": [
          {
            "jwt_access_token": []
          }
        ],
        "summary": "Request activation link",
        "tags": [
          "Authentication operations"
        ]
      }
    },
    "/auth/devices": {
      "delete": {
        "parameters": [],
        "responses": {
          "204": {
            "description": "Logged out from all devices"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          }
        },
        "security": [
          {
            "jwt_access_token": []
          }
        ],
        "summary": "Log out from all devices",
        "tags": [
          "Authentication operations"
        ]
      },
      "get": {
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Device"
                  },
                  "type": "array"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          }
        },
        "security": [
          {
            "jwt_access_token": []
          }
        ],
        "summary": "Get all active devices",
        "tags": [
          "Authentication operations"
        ]
      }
    },
    "/auth/devices/{device_id}": {
      "delete": {
        "parameters": [
          {
            "in": "path",
            "name": "device_id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Password"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Device logged out"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Invalid password"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Device not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "security": [
          {
            "jwt_access_token": []
          }
        ],
        "summary": "Log out from the device",
        "tags": [
          "Authentication operations"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Login"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessToken"
                }
              }
            },
            "description": "User logged in"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Invalid email or password"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Login user. Set refresh token cookie.",
        "tags": [
          "Authentication operations"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          }
        },
        "security": [
          {
            "jwt_access_token": []
          }
        ],
        "summary": "Log out from the current device. Unset refresh token cookie.",
        "tags": [
          "Authentication operations"
        ]
      }
    },
    "/auth/password": {
      "put": {
        "parameters": [],
        "responses": {
          "204": {
            "description": "Password changed"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Invalid old or new password"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          }
        },
        "security": [
          {
            "jwt_access_token": []
          }
        ],
        "summary": "Change password",
        "tags": [
          "Authentication operations"
        ]
      }
    },
    "/auth/password/reset": {
      "post": {
        "parameters": [],
        "responses": {
          "204": {
            "description": "Password changed"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Invalid secret code or new password"
          }
        },
        "summary": "Reset password using a secret code from email and pass a new password",
        "tags": [
          "Authentication operations"
        ]
      }
    },
...
    "/auth/refresh": {
      "post": {
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessToken"
                }
              }
            },
            "description": "Access token refreshed"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          }
        },
        "security": [
          {
            "jwt_refresh_token": []
          },
          {
            "csrf_refresh_token": []
          }
        ],
        "summary": "Refresh access token",
        "tags": [
          "Authentication operations"
        ]
      }
    },
    "/auth/register": {
      "post": {
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Register"
              }
            }
          }
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccessToken"
                }
              }
            },
            "description": "User registered",
            "headers": {
              "location": {
                "description": "URL of the created resource",
                "required": false,
                "schema": {
                  "format": "url",
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Invalid email or locality_id"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Email already used"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            },
            "description": "Validation error"
          }
        },
        "summary": "Register a new user. Set refresh token cookie. Send the activation link.",
        "tags": [
          "Authentication operations"
        ]
      }
    },
    "/auth/whoami": {
      "get": {
        "parameters": [],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WhoAmI"
                }
              }
            },
            "description": "Successful response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPError"
                }
              }
            },
            "description": "Authentication error"
          }
        },
        "security": [
          {
            "jwt_access_token": []
          }
        ],
        "summary": "Get current user info",
        "tags": [
          "Authentication operations"
        ]
      }
    },
...
  },
...
}